requirements.txt
asgiref==3.2.10
Django==3.0.8
djangorestframework==3.11.0
pytz==2020.1
sqlparse==0.3.1
django-allauth==0.39.1
django-audit-log==0.7.0
django-auditlog==0.4.3
django-cors-headers==3.5.0
django-crispy-forms==1.7.2
django-mysql==3.3.0
django-jsonfield==1.0.1
dj-database-url==0.5.0
pillow==2.2.1
psycopg2==2.8.4
pycparser==2.17
urllib3==1.24.2